added :indexCopy

+++ lib/THCl/THClTensorIndex.cl
@@ -40,10 +40,10 @@ kernel void THClTensor_kernel_indexFill(
 }
 
 kernel void THClTensor_kernel_indexCopy(
-   global float *res_data, long res_offset, 
-   global float *src_data, long src_offset,
-   global long* res_stride, global float *index_data, long index_offset,
-   long res_nDim, int dim, long idx_size, long src_size, long size_dim
+   global float *res_data, int res_offset, 
+   global float *src_data, int src_offset,
+   global int* res_stride, global float *index_data, int index_offset,
+   int res_nDim, int dim, int idx_size, int src_size, int size_dim
 )
 {
   int thread_idx = get_group_id(0) * get_local_size(0) * get_local_size(1) + get_local_id(1) * get_local_size(0) + get_local_id(0);
